requirements.txt
 fastapi
 uvicorn
 firebase-admin
 python-dotenv
 psycopg2